Add gitcd fish function
This allows me to drop into any repository from anywhere. Since I track
the gitlab group hierarchy, adding to CDPATH doesn't scale as I would
need to keep adding paths. The drawback to this is there are a handful
of repos that share the same name. This causes some to be unreachable. I
also would like this to have autocomplete, but hey maybe that's for a
rainy day.