Prevent tmux from starting in vscode
I don't use the vscode terminal much, but when I do (starting debugging, for
example), it will attach to the same tmux session. This results in odd behavior
where my external connection to the tmux session is adjusted to the size of the
terminal panel in vscode. Not ideal. If I ever decide I need a second tmux
session, I can add it later, but I don't think I need one for now.